Move `ChunkWitness` out of `pieces` module
